pyproject.toml
[tool.poetry]
name = "mpttdemo"
version = "0.1.0"
description = ""
authors = ["rhami223 <rhami223@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
